Refactor TinkerEngine to use backend architecture
Introduces a clean separation between engine (orchestration) and backend (computation):

**New files in `backends/`:**
- `backend.py`: AbstractBackend interface defining the contract
- `native.py`: NativeBackend implementation (extracted from engine.py)
- `utils.py`: Shared utilities (log_timing, pad, pad_batch)
- `__init__.py`: Module exports

**Engine responsibilities (engine.py):**
- Database operations (futures, checkpoints)
- Request validation (`_filter_valid_requests`)
- Data extraction from requests (`_prepare_model_pass_batch`, `_prepare_sample_batch`)
- File I/O (checkpoint download/upload)
- Orchestration of batch processing

**Backend responsibilities (native.py):**
- Model initialization and state management
- JAX/Flax computation (forward, backward, gradient accumulation)
- Optimizer creation and updates
- Checkpoint data extraction/insertion

**New types in types.py:**
- `PreparedModelPassBatch`: Batch data for forward/backward ops
- `PreparedSampleBatch`: Batch data for sampling ops

This is a purely structural refactor - no functional changes to computation logic.
